to move directories gentx, load, and run under lib into lib/cmd
