+++ lib/cmd/run/generator.go
